+++ R/np.distribution.R
@@ -1,5 +1,7 @@
 npudist <-
   function(bws, ...){
+    mc <- match.call(expand.dots = FALSE)
+    .np_validate_public_dots(mc[["..."]], "npudist")
     args <- list(...)
 
     if (!missing(bws)){
@@ -298,6 +300,7 @@ npudist.default <- function(bws, tdat, ...){
   if(any(m.txy > 0)) {
     names(sc.bw)[m.txy] <- nstxy[m.txy > 0]
   }
+  sc.bw <- .np_public_dots_filter_call(sc.bw, "npudistbw")
     
   tbw <- if (!has.explicit.bws) {
     .np_progress_select_bandwidth_enhanced(