+++ R/np.plregression.R
@@ -1,5 +1,7 @@
 npplreg <-
   function(bws, ...){
+    mc <- match.call(expand.dots = FALSE)
+    .np_validate_public_dots(mc[["..."]], "npplreg")
     args <- list(...)
 
     if (!missing(bws)){
@@ -749,6 +751,7 @@ npplreg.default <- function(bws, txdat, tydat, tzdat, nomad = FALSE, ...) {
   if(any(m.txy > 0)) {
     names(sc.bw)[m.txy] <- nstxy[m.txy > 0]
   }
+  sc.bw <- .np_public_dots_filter_call(sc.bw, "npplregbw")
     
   use.outer.bandwidth.progress <- !.np_bw_call_uses_nomad_degree_search(
     sc.bw,